Updated typos Created currentassetsinvestments package and moved CurrentAssetsInvestment model to it
