+++ src/main/java/uk/gov/companieshouse/web/accounts/service/smallfull/CurrentAssetsInvestmentsService.java
@@ -1,7 +1,7 @@
 package uk.gov.companieshouse.web.accounts.service.smallfull;
 
 import uk.gov.companieshouse.web.accounts.exception.ServiceException;
-import uk.gov.companieshouse.web.accounts.model.smallfull.notes.CurrentAssetsInvestments;
+import uk.gov.companieshouse.web.accounts.model.smallfull.notes.currentassetsinvestments.CurrentAssetsInvestments;
 import uk.gov.companieshouse.web.accounts.validation.ValidationError;
 
 import java.util.List;
