+++ src/main/java/uk/gov/companieshouse/web/accounts/transformer/smallfull/CurrentAssetsInvestmentsTransformer.java
@@ -1,7 +1,7 @@
 package uk.gov.companieshouse.web.accounts.transformer.smallfull;
 
 import uk.gov.companieshouse.api.model.accounts.smallfull.currentassetsinvestments.CurrentAssetsInvestmentsApi;
-import uk.gov.companieshouse.web.accounts.model.smallfull.notes.CurrentAssetsInvestments;
+import uk.gov.companieshouse.web.accounts.model.smallfull.notes.currentassetsinvestments.CurrentAssetsInvestments;
 
 public interface CurrentAssetsInvestmentsTransformer {
